# app/api/routes.py
from flask import Blueprint, request, jsonify, current_app
from ..core.scanner import process_command, scan_with_nmap, scan_with_zap
from ..updater import update_program
from datetime import datetime
import logging
from functools import wraps
import jwt
from .payload_detector import PayloadDetector

main = Blueprint('main', __name__)

# Authentication decorator
def token_required(f):
    @wraps(f)
    def decorated(*args, **kwargs):
        token = request.headers.get('X-API-Token')
        if not token:
            return jsonify({'error': 'Token is missing'}), 401
        try:
            jwt.decode(token, current_app.config['SECRET_KEY'], algorithms=["HS256"])
        except:
            return jsonify({'error': 'Invalid token'}), 401
        return f(*args, **kwargs)
    return decorated

@main.route('/')
def index():
    return "AI Vulnerability Scanner Assistant"

@main.route('/api/scan', methods=['POST'])
@token_required
def scan_endpoint():
    try:
        data = request.get_json()
        target = data.get('target')
        scan_type = data.get('scan_type', 'full')
        
        if not target:
            return jsonify({'error': 'Target URL/IP is required'}), 400
            
        timestamp = datetime.now().isoformat()
        results = {
            'timestamp': timestamp,
            'target': target,
            'scan_type': scan_type,
            'results': {}
        }
        
        if scan_type in ['full', 'nmap']:
            results['results']['nmap'] = scan_with_nmap(target, timestamp)
            
        if scan_type in ['full', 'zap']:
            results['results']['zap'] = scan_with_zap(target, timestamp)
            
        # Log scan activity
        logging.info(f"Scan completed for {target} using {scan_type} scan")
        
        return jsonify(results)
        
    except Exception as e:
        logging.error(f"Scan error: {str(e)}")
        return jsonify({'error': str(e)}), 500

@main.route('/api/analyze-payload', methods=['POST'])
@token_required
def analyze_payload():
    try:
        data = request.get_json()
        content = data.get('content')
        
        if not content:
            return jsonify({'error': 'Payload content is required'}), 400
            
        detector = PayloadDetector()
        detections = detector.analyze_payload(content)
        report = detector.generate_report(detections)
        
        return jsonify(report)
        
    except Exception as e:
        logging.error(f"Payload analysis error: {str(e)}")
        return jsonify({'error': str(e)}), 500

@main.route('/api/scan-history', methods=['GET'])
@token_required
def scan_history():
    try:
        # Get parameters for filtering
        target = request.args.get('target')
        start_date = request.args.get('start_date')
        end_date = request.args.get('end_date')
        
        # Here you would implement database querying logic
        # This is a placeholder for demonstration
        history = {
            'scans': [
                {
                    'id': 'scan123',
                    'timestamp': '2024-01-25T10:00:00',
                    'target': 'example.com',
                    'type': 'full',
                    'status': 'completed'
                }
            ]
        }
        
        return jsonify(history)
        
    except Exception as e:
        logging.error(f"History retrieval error: {str(e)}")
        return jsonify({'error': str(e)}), 500

@main.route('/api/report', methods=['GET'])
@token_required
def get_report():
    try:
        scan_id = request.args.get('scan_id')
        if not scan_id:
            return jsonify({'error': 'Scan ID is required'}), 400
            
        # Here you would implement report generation logic
        # This is a placeholder for demonstration
        report = {
            'scan_id': scan_id,
            'generated_at': datetime.now().isoformat(),
            'summary': {
                'total_vulnerabilities': 5,
                'high_severity': 2,
                'medium_severity': 2,
                'low_severity': 1
            },
            'details': []
        }
        
        return jsonify(report)
        
    except Exception as e:
        logging.error(f"Report generation error: {str(e)}")
        return jsonify({'error': str(e)}), 500

@main.route('/api/update', methods=['POST'])
@token_required
def update():
    try:
        response = update_program()
        return jsonify(response)
    except Exception as e:
        logging.error(f"Update error: {str(e)}")
        return jsonify({'error': str(e)}), 500

# Error handlers
@main.errorhandler(404)
def not_found(error):
    return jsonify({'error': 'Not found'}), 404

@main.errorhandler(500)
def internal_error(error):
    return jsonify({'error': 'Internal server error'}), 500