# app/models/report.py
# app/models/report.py

from datetime import datetime
import uuid
from app import db

class Report(db.Model):
    """Report model for storing generated reports"""
    __tablename__ = 'reports'

    id = db.Column(db.String(36), primary_key=True, default=lambda: str(uuid.uuid4()))
    scan_id = db.Column(db.String(36), db.ForeignKey('scans.id'), nullable=False)
    user_id = db.Column(db.String(36), db.ForeignKey('users.id'), nullable=False)
    title = db.Column(db.String(200), nullable=False)
    report_type = db.Column(db.String(20), nullable=False)  # e.g., 'detailed', 'executive', 'compliance'
    format = db.Column(db.String(10), nullable=False)  # e.g., 'pdf', 'html', 'json'
    status = db.Column(db.String(20), default='pending')  # pending, generating, completed, failed
    file_path = db.Column(db.String(255))
    file_size = db.Column(db.Integer)
    template_used = db.Column(db.String(50))
    configuration = db.Column(db.JSON)
    generated_at = db.Column(db.DateTime)
    expires_at = db.Column(db.DateTime)
    created_at = db.Column(db.DateTime, default=datetime.utcnow)
    updated_at = db.Column(db.DateTime, default=datetime.utcnow, onupdate=datetime.utcnow)

    # Relationships
    sections = db.relationship('ReportSection', backref='report', lazy='dynamic')
    comments = db.relationship('ReportComment', backref='report', lazy='dynamic')

    def start_generation(self):
        """Start report generation process"""
        self.status = 'generating'
        db.session.commit()

    def complete_generation(self, file_path, file_size):
        """Mark report generation as complete"""
        self.status = 'completed'
        self.file_path = file_path
        self.file_size = file_size
        self.generated_at = datetime.utcnow()
        db.session.commit()

    def fail_generation(self, error_message):
        """Mark report generation as failed"""
        self.status = 'failed'
        self.add_comment('system', f'Generation failed: {error_message}')
        db.session.commit()

    def is_expired(self):
        """Check if report has expired"""
        if not self.expires_at:
            return False
        return datetime.utcnow() > self.expires_at

    def add_section(self, title, content, order=None):
        """Add section to report"""
        section = ReportSection(
            report_id=self.id,
            title=title,
            content=content,
            order=order or self.sections.count() + 1
        )
        db.session.add(section)
        db.session.commit()
        return section

    def add_comment(self, comment_type, content):
        """Add comment to report"""
        comment = ReportComment(
            report_id=self.id,
            comment_type=comment_type,
            content=content
        )
        db.session.add(comment)
        db.session.commit()
        return comment

    def to_dict(self):
        """Convert report to dictionary"""
        return {
            'id': self.id,
            'scan_id': self.scan_id,
            'user_id': self.user_id,
            'title': self.title,
            'report_type': self.report_type,
            'format': self.format,
            'status': self.status,
            'file_path': self.file_path,
            'file_size': self.file_size,
            'template_used': self.template_used,
            'configuration': self.configuration,
            'generated_at': self.generated_at.isoformat() if self.generated_at else None,
            'expires_at': self.expires_at.isoformat() if self.expires_at else None,
            'created_at': self.created_at.isoformat(),
            'updated_at': self.updated_at.isoformat()
        }

class ReportSection(db.Model):
    """Model for storing report sections"""
    __tablename__ = 'report_sections'

    id = db.Column(db.String(36), primary_key=True, default=lambda: str(uuid.uuid4()))
    report_id = db.Column(db.String(36), db.ForeignKey('reports.id'), nullable=False)
    title = db.Column(db.String(200), nullable=False)
    content = db.Column(db.Text, nullable=False)
    order = db.Column(db.Integer, nullable=False)
    created_at = db.Column(db.DateTime, default=datetime.utcnow)
    updated_at = db.Column(db.DateTime, default=datetime.utcnow, onupdate=datetime.utcnow)

    def to_dict(self):
        """Convert section to dictionary"""
        return {
            'id': self.id,
            'report_id': self.report_id,
            'title': self.title,
            'content': self.content,
            'order': self.order,
            'created_at': self.created_at.isoformat(),
            'updated_at': self.updated_at.isoformat()
        }

class ReportComment(db.Model):
    """Model for storing comments on reports"""
    __tablename__ = 'report_comments'

    id = db.Column(db.String(36), primary_key=True, default=lambda: str(uuid.uuid4()))
    report_id = db.Column(db.String(36), db.ForeignKey('reports.id'), nullable=False)
    user_id = db.Column(db.String(36), db.ForeignKey('users.id'))
    comment_type = db.Column(db.String(20), nullable=False)  # e.g., 'user', 'system', 'review'
    content = db.Column(db.Text, nullable=False)
    created_at = db.Column(db.DateTime, default=datetime.utcnow)
    updated_at = db.Column(db.DateTime, default=datetime.utcnow, onupdate=datetime.utcnow)

    # Relationship
    author = db.relationship('User', backref='report_comments')

    def to_dict(self):
        """Convert comment to dictionary"""
        return {
            'id': self.id,
            'report_id': self.report_id,
            'user_id': self.user_id,
            'comment_type': self.comment_type,
            'content': self.content,
            'created_at': self.created_at.isoformat(),
            'updated_at': self.updated_at.isoformat()
        }

class ReportTemplate(db.Model):
    """Model for storing report templates"""
    __tablename__ = 'report_templates'

    id = db.Column(db.String(36), primary_key=True, default=lambda: str(uuid.uuid4()))
    name = db.Column(db.String(100), nullable=False, unique=True)
    description = db.Column(db.Text)
    template_type = db.Column(db.String(20), nullable=False)  # e.g., 'detailed', 'executive', 'compliance'
    format = db.Column(db.String(10), nullable=False)  # e.g., 'pdf', 'html', 'json'
    content = db.Column(db.Text, nullable=False)
    css_content = db.Column(db.Text)
    version = db.Column(db.String(10))
    is_active = db.Column(db.Boolean, default=True)
    created_by = db.Column(db.String(36), db.ForeignKey('users.id'))
    created_at = db.Column(db.DateTime, default=datetime.utcnow)
    updated_at = db.Column(db.DateTime, default=datetime.utcnow, onupdate=datetime.utcnow)

    def to_dict(self):
        """Convert template to dictionary"""
        return {
            'id': self.id,
            'name': self.name,
            'description': self.description,
            'template_type': self.template_type,
            'format': self.format,
            'version': self.version,
            'is_active': self.is_active,
            'created_by': self.created_by,
            'created_at': self.created_at.isoformat(),
            'updated_at': self.updated_at.isoformat()
        }