# app/models/scan.py
# app/models/scan.py

from datetime import datetime
import uuid
import json
from app import db

class Scan(db.Model):
    """Scan model for storing scan information"""
    __tablename__ = 'scans'

    id = db.Column(db.String(36), primary_key=True, default=lambda: str(uuid.uuid4()))
    user_id = db.Column(db.String(36), db.ForeignKey('users.id'), nullable=False)
    target = db.Column(db.String(255), nullable=False)
    scan_type = db.Column(db.String(20), nullable=False)
    status = db.Column(db.String(20), default='pending')
    progress = db.Column(db.Integer, default=0)
    start_time = db.Column(db.DateTime)
    end_time = db.Column(db.DateTime)
    configuration = db.Column(db.JSON)
    created_at = db.Column(db.DateTime, default=datetime.utcnow)
    updated_at = db.Column(db.DateTime, default=datetime.utcnow, onupdate=datetime.utcnow)

    # Relationships
    results = db.relationship('ScanResult', backref='scan', lazy='dynamic')
    vulnerabilities = db.relationship('Vulnerability', backref='scan', lazy='dynamic')
    reports = db.relationship('Report', backref='scan', lazy='dynamic')

    def start(self):
        """Start the scan"""
        self.status = 'running'
        self.start_time = datetime.utcnow()
        self.progress = 0
        db.session.commit()

    def complete(self):
        """Complete the scan"""
        self.status = 'completed'
        self.end_time = datetime.utcnow()
        self.progress = 100
        db.session.commit()

    def fail(self, error_message):
        """Mark scan as failed"""
        self.status = 'failed'
        self.end_time = datetime.utcnow()
        self.add_result('error', error_message)
        db.session.commit()

    def update_progress(self, progress):
        """Update scan progress"""
        self.progress = progress
        db.session.commit()

    def add_result(self, result_type, data):
        """Add scan result"""
        result = ScanResult(
            scan_id=self.id,
            result_type=result_type,
            data=data
        )
        db.session.add(result)
        db.session.commit()
        return result

    def add_vulnerability(self, title, severity, description, **kwargs):
        """Add vulnerability finding"""
        vuln = Vulnerability(
            scan_id=self.id,
            title=title,
            severity=severity,
            description=description,
            **kwargs
        )
        db.session.add(vuln)
        db.session.commit()
        return vuln

    def to_dict(self):
        """Convert scan object to dictionary"""
        return {
            'id': self.id,
            'user_id': self.user_id,
            'target': self.target,
            'scan_type': self.scan_type,
            'status': self.status,
            'progress': self.progress,
            'start_time': self.start_time.isoformat() if self.start_time else None,
            'end_time': self.end_time.isoformat() if self.end_time else None,
            'configuration': self.configuration,
            'created_at': self.created_at.isoformat(),
            'updated_at': self.updated_at.isoformat()
        }

class ScanResult(db.Model):
    """Scan result model for storing detailed scan results"""
    __tablename__ = 'scan_results'

    id = db.Column(db.String(36), primary_key=True, default=lambda: str(uuid.uuid4()))
    scan_id = db.Column(db.String(36), db.ForeignKey('scans.id'), nullable=False)
    result_type = db.Column(db.String(50), nullable=False)
    data = db.Column(db.JSON)
    created_at = db.Column(db.DateTime, default=datetime.utcnow)

    def to_dict(self):
        """Convert scan result to dictionary"""
        return {
            'id': self.id,
            'scan_id': self.scan_id,
            'result_type': self.result_type,
            'data': self.data,
            'created_at': self.created_at.isoformat()
        }

class Vulnerability(db.Model):
    """Vulnerability model for storing vulnerability findings"""
    __tablename__ = 'vulnerabilities'

    id = db.Column(db.String(36), primary_key=True, default=lambda: str(uuid.uuid4()))
    scan_id = db.Column(db.String(36), db.ForeignKey('scans.id'), nullable=False)
    title = db.Column(db.String(200), nullable=False)
    severity = db.Column(db.String(20), nullable=False)
    description = db.Column(db.Text, nullable=False)
    location = db.Column(db.String(255))
    vulnerability_type = db.Column(db.String(50))
    proof_of_concept = db.Column(db.Text)
    remediation = db.Column(db.Text)
    cvss_score = db.Column(db.Float)
    cve_ids = db.Column(db.JSON)
    tags = db.Column(db.JSON)
    status = db.Column(db.String(20), default='open')
    discovered_at = db.Column(db.DateTime, default=datetime.utcnow)
    updated_at = db.Column(db.DateTime, default=datetime.utcnow, onupdate=datetime.utcnow)
    resolved_at = db.Column(db.DateTime)

    def resolve(self, resolution_notes=None):
        """Mark vulnerability as resolved"""
        self.status = 'resolved'
        self.resolved_at = datetime.utcnow()
        if resolution_notes:
            self.add_note('resolution', resolution_notes)
        db.session.commit()

    def reopen(self, reason=None):
        """Reopen resolved vulnerability"""
        self.status = 'open'
        self.resolved_at = None
        if reason:
            self.add_note('reopened', reason)
        db.session.commit()

    def add_note(self, note_type, content):
        """Add note to vulnerability"""
        note = VulnerabilityNote(
            vulnerability_id=self.id,
            note_type=note_type,
            content=content
        )
        db.session.add(note)
        db.session.commit()
        return note

    def to_dict(self):
        """Convert vulnerability to dictionary"""
        return {
            'id': self.id,
            'scan_id': self.scan_id,
            'title': self.title,
            'severity': self.severity,
            'description': self.description,
            'location': self.location,
            'vulnerability_type': self.vulnerability_type,
            'proof_of_concept': self.proof_of_concept,
            'remediation': self.remediation,
            'cvss_score': self.cvss_score,
            'cve_ids': self.cve_ids,
            'tags': self.tags,
            'status': self.status,
            'discovered_at': self.discovered_at.isoformat(),
            'updated_at': self.updated_at.isoformat(),
            'resolved_at': self.resolved_at.isoformat() if self.resolved_at else None
        }


class VulnerabilityNote(db.Model):
    """Model for storing notes about vulnerabilities"""
    __tablename__ = 'vulnerability_notes'

    id = db.Column(db.String(36), primary_key=True, default=lambda: str(uuid.uuid4()))
    vulnerability_id = db.Column(db.String(36), db.ForeignKey('vulnerabilities.id'), nullable=False)
    note_type = db.Column(db.String(20), nullable=False)  # e.g., 'comment', 'resolution', 'reopened'
    content = db.Column(db.Text, nullable=False)
    created_by = db.Column(db.String(36), db.ForeignKey('users.id'))
    created_at = db.Column(db.DateTime, default=datetime.utcnow)
    updated_at = db.Column(db.DateTime, default=datetime.utcnow, onupdate=datetime.utcnow)

    # Relationships
    vulnerability = db.relationship('Vulnerability', backref='notes')
    author = db.relationship('User', backref='vulnerability_notes')

    def to_dict(self):
        """Convert note to dictionary"""
        return {
            'id': self.id,
            'vulnerability_id': self.vulnerability_id,
            'note_type': self.note_type,
            'content': self.content,
            'created_by': self.created_by,
            'created_at': self.created_at.isoformat(),
            'updated_at': self.updated_at.isoformat()
        }