# app/static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "AI Vulnerability Scanner Assistant API",
    "version": "1.1.0",
    "description": "RESTful API for the AI-powered Vulnerability Scanner Assistant",
    "contact": {
      "name": "Security Team",
      "email": "security@example.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "api.vulnscan.example.com",
  "basePath": "/api/v1",
  "schemes": ["https"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "securityDefinitions": {
    "ApiKeyAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "X-API-Key"
    },
    "BearerAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "description": "JWT token prefixed with 'Bearer '"
    }
  },
  "security": [
    {
      "ApiKeyAuth": [],
      "BearerAuth": []
    }
  ],
  "paths": {
    "/scan": {
      "post": {
        "tags": ["Scanning"],
        "summary": "Initiate a new security scan",
        "description": "Start a new vulnerability scan on the specified target",
        "parameters": [
          {
            "in": "body",
            "name": "scanRequest",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ScanRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Scan initiated successfully",
            "schema": {
              "$ref": "#/definitions/ScanResponse"
            }
          },
          "400": {
            "description": "Invalid request parameters",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized access",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "429": {
            "description": "Too many requests",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/scan/{scanId}": {
      "get": {
        "tags": ["Scanning"],
        "summary": "Get scan results",
        "description": "Retrieve the results of a specific scan",
        "parameters": [
          {
            "name": "scanId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Scan results retrieved successfully",
            "schema": {
              "$ref": "#/definitions/ScanResult"
            }
          },
          "404": {
            "description": "Scan not found",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/analyze-payload": {
      "post": {
        "tags": ["Analysis"],
        "summary": "Analyze payload for vulnerabilities",
        "description": "Analyze provided content for potential security vulnerabilities",
        "parameters": [
          {
            "in": "body",
            "name": "payload",
            "required": true,
            "schema": {
              "$ref": "#/definitions/PayloadAnalysisRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Analysis completed successfully",
            "schema": {
              "$ref": "#/definitions/PayloadAnalysisResponse"
            }
          }
        }
      }
    },
    "/report/{scanId}": {
      "get": {
        "tags": ["Reporting"],
        "summary": "Generate scan report",
        "description": "Generate a detailed report for a specific scan",
        "parameters": [
          {
            "name": "scanId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "format",
            "in": "query",
            "type": "string",
            "enum": ["json", "pdf", "html"],
            "default": "json"
          }
        ],
        "responses": {
          "200": {
            "description": "Report generated successfully",
            "schema": {
              "$ref": "#/definitions/Report"
            }
          }
        }
      }
    },
    "/update": {
      "post": {
        "tags": ["System"],
        "summary": "Update system",
        "description": "Update the vulnerability scanner system",
        "responses": {
          "200": {
            "description": "Update completed successfully",
            "schema": {
              "$ref": "#/definitions/UpdateResponse"
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "tags": ["System"],
        "summary": "System health check",
        "description": "Check the health status of the system",
        "security": [],
        "responses": {
          "200": {
            "description": "System health status",
            "schema": {
              "$ref": "#/definitions/HealthResponse"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "ScanRequest": {
      "type": "object",
      "required": ["target"],
      "properties": {
        "target": {
          "type": "string",
          "description": "URL or IP address to scan"
        },
        "scanType": {
          "type": "string",
          "enum": ["full", "quick", "custom"],
          "default": "full"
        },
        "options": {
          "type": "object",
          "properties": {
            "portScan": {
              "type": "boolean",
              "default": true
            },
            "vulnerabilityScan": {
              "type": "boolean",
              "default": true
            },
            "sslScan": {
              "type": "boolean",
              "default": true
            }
          }
        }
      }
    },
    "ScanResponse": {
      "type": "object",
      "properties": {
        "scanId": {
          "type": "string"
        },
        "status": {
          "type": "string",
          "enum": ["initiated", "in_progress", "completed", "failed"]
        },
        "estimatedTime": {
          "type": "integer",
          "description": "Estimated time in seconds"
        }
      }
    },
    "ScanResult": {
      "type": "object",
      "properties": {
        "scanId": {
          "type": "string"
        },
        "target": {
          "type": "string"
        },
        "timestamp": {
          "type": "string",
          "format": "date-time"
        },
        "vulnerabilities": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Vulnerability"
          }
        },
        "summary": {
          "$ref": "#/definitions/ScanSummary"
        }
      }
    },
    "Vulnerability": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "type": {
          "type": "string"
        },
        "severity": {
          "type": "string",
          "enum": ["LOW", "MEDIUM", "HIGH", "CRITICAL"]
        },
        "description": {
          "type": "string"
        },
        "impact": {
          "type": "string"
        },
        "recommendation": {
          "type": "string"
        },
        "references": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "ScanSummary": {
      "type": "object",
      "properties": {
        "totalVulnerabilities": {
          "type": "integer"
        },
        "criticalCount": {
          "type": "integer"
        },
        "highCount": {
          "type": "integer"
        },
        "mediumCount": {
          "type": "integer"
        },
        "lowCount": {
          "type": "integer"
        }
      }
    },
    "PayloadAnalysisRequest": {
      "type": "object",
      "required": ["content"],
      "properties": {
        "content": {
          "type": "string",
          "description": "Content to analyze"
        },
        "contentType": {
          "type": "string",
          "enum": ["text", "json", "xml", "binary"]
        }
      }
    },
    "PayloadAnalysisResponse": {
      "type": "object",
      "properties": {
        "detections": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Detection"
          }
        },
        "riskLevel": {
          "type": "string",
          "enum": ["LOW", "MEDIUM", "HIGH", "CRITICAL"]
        }
      }
    },
    "Detection": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string"
        },
        "signature": {
          "type": "string"
        },
        "confidence": {
          "type": "number",
          "format": "float",
          "minimum": 0,
          "maximum": 1
        },
        "location": {
          "type": "string"
        }
      }
    },
    "Report": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "generatedAt": {
          "type": "string",
          "format": "date-time"
        },
        "summary": {
          "$ref": "#/definitions/ScanSummary"
        },
        "details": {
          "type": "object"
        }
      }
    },
    "UpdateResponse": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "enum": ["success", "error"]
        },
        "message": {
          "type": "string"
        },
        "version": {
          "type": "string"
        }
      }
    },
    "HealthResponse": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "enum": ["healthy", "unhealthy"]
        },
        "version": {
          "type": "string"
        },
        "timestamp": {
          "type": "string",
          "format": "date-time"
        },
        "components": {
          "type": "object",
          "properties": {
            "scanner": {
              "type": "string"
            },
            "database": {
              "type": "string"
            }
          }
        }
      }
    },
    "ErrorResponse": {
      "type": "object",
      "properties": {
        "error": {
          "type": "string"
        },
        "message": {
          "type": "string"
        },
        "code": {
          "type": "integer"
        }
      }
    }
  },
  "tags": [
    {
      "name": "Scanning",
      "description": "Vulnerability scanning operations"
    },
    {
      "name": "Analysis",
      "description": "Payload analysis operations"
    },
    {
      "name": "Reporting",
      "description": "Report generation operations"
    },
    {
      "name": "System",
      "description": "System management operations"
    }
  ]
}